Refactored schedules controller actions to improve parameter handling, JSON response, and API data retrieval

+++ app/controllers/collected_plants_controller.rb
@@ -2,49 +2,50 @@ require 'http'
 class CollectedPlantsController < ApplicationController
   before_action :authenticate_user
 
-
-  def index 
+  def index
     @collected_plants = current_user.collected_plants.order(created_at: :desc)
-    render :index
+    collected_plants_data = []
+
+    @collected_plants.each do |collected_plant|
+      plant_data = retrieve_plant_data(collected_plant[:plant_id])
+      next unless plant_data
+
+      schedule_data = collected_plant.schedule ? schedule_data(collected_plant.schedule) : 'No schedule available'
+
+      collected_plant_data = {
+        'collected_plant' => collected_plant,
+        'plant_data' => plant_data,
+        'schedule' => schedule_data
+      }
+
+      collected_plants_data << collected_plant_data
+    end
+
+    render json: collected_plants_data
   end
 
   def show
     @collected_plant = CollectedPlant.find(params[:id])
+    plant_data = retrieve_plant_data(@collected_plant[:plant_id])
+    schedule_data = @collected_plant.schedule ? schedule_data(@collected_plant.schedule) : 'No schedule available'
 
-    response = api_get_request("https://house-plants2.p.rapidapi.com/id/#{@collected_plant[:plant_id]}")
-
-    if response.code == 200
-      plant_data = JSON.parse(response.body)
-      @plant_data = {
-        'id' => plant_data['id'],
-        'common_name' => plant_data['Common name (fr.)'] || 'Unknown',
-        'latin_name' => plant_data['Latin name'] || 'Unknown',
-        'img' => plant_data['Img'] || 'Unknown',
-        'watering' => plant_data['Watering'] || 'Unknown',
-        'light_ideal' => plant_data['Light ideal'] || 'Unknown',
-        'light_tolerated' => plant_data['Light tolered'] || 'Unknown',
-        'climate' => plant_data['Climat'] || 'Unknown',
-        'category' => plant_data['Categories'] || 'Unknown',
-        'url' => plant_data['Url'] || 'Resource not found'
+    if plant_data
+      render json: {
+        collected_plant: @collected_plant,
+        plant_data: plant_data,
+        schedule: schedule_data
       }
-
-      render json: { collected_plant: @collected_plant, plant_data: @plant_data }
     else
-      render json: { error: "Failed to retrieve plant data. Error code: #{response.code}" }, status: :unprocessable_entity
+      render json: { error: "Failed to retrieve plant data." }, status: :unprocessable_entity
     end
   end
 
   def create
     plant_id = params[:plant_id]
-    user_id = current_user.id
+    @user_id = current_user.id
     @plant_data = retrieve_plant_data(plant_id)
 
-    @collected_plant = CollectedPlant.new(
-      plant_id: plant_id,
-      user_id: user_id,
-      nickname: params[:nickname],
-      notes: params[:notes]
-    )
+    @collected_plant = CollectedPlant.new(collected_plant_params)
 
     if @collected_plant.save
       redirect_to @collected_plant, notice: "Plant saved to collection successfully."
@@ -55,7 +56,7 @@ class CollectedPlantsController < ApplicationController
   end
 
   def update
-    @collected_plant = CollectedPlant.find_by(id: params[:id])
+    @collected_plant = CollectedPlant.find_by(id: params[:collected_plant_id])
   
     if @collected_plant
       if @collected_plant.update(collected_plant_params)
@@ -71,10 +72,14 @@ class CollectedPlantsController < ApplicationController
   def destroy
     @collected_plant = CollectedPlant.find_by(id: params[:id])
 
-    if @collected_plant.destroy
-      render json: { message: "Collected plant destroyed successfully" }
+    if @collected_plant
+      if @collected_plant.destroy
+        render json: { message: "Collected plant destroyed successfully" }
+      else
+        render json: { errors: @collected_plant.errors.full_messages }, status: :unprocessable_entity
+      end
     else
-      render json: { message: "Collected Plant not found" }, status: :not_found
+      render json: { errors: ['Collected Plant not found'] }, status: :not_found
     end
   end
 
@@ -87,6 +92,15 @@ class CollectedPlantsController < ApplicationController
   def retrieve_plant_data(plant_id)
     url = "https://house-plants2.p.rapidapi.com/id/#{plant_id}"
     response = api_get_request(url)
-    JSON.parse(response.body)
+    response.code == 200 ? JSON.parse(response.body) : nil
+  end
+
+  def schedule_data(schedule)
+    {
+      'id' => schedule.id,
+      'collected_plant_id' => schedule.collected_plant_id,
+      'watering_start_date' => schedule.watering_start_date,
+      'days_to_water' => schedule.days_to_water
+    }
   end
 end
